We're in FULL BETA
This is the final BETA release of SImplistic Design (not including testing releases). The downloads have not been updated, but you can download the main styles and scripts. In this release, a lot was added, including buttons, a side nav dropdown option, a ripple effect for the buttons (works on mobile, too!). Added a dialog, and FAB button. I added some "syntax" pages, but those are not finished although they are accessible from the main website. Stay tuned for version 1.0 of Simplistic Design! - Isaac Maddox, Simplistic Design, founder and developer