A LOT HAPPENED
I fixed all of the main pages for Simplistic Design so that they match the current state of it. I added a blog page, which will contain release info and development info as well. I added a feature to the selects, so now if the menu goes over the top of the screen, the overflow is set to auto and the user can scroll through the options instead of having to scroll through the actual page to see them. Border on bottom of inputs, textarea, and select is now 2px instead of 1. Like I said, A LOT HAPPENED.